backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.8.2
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.4
python-dotenv==1.0.1
sqlalchemy==2.0.35
dashscope==1.19.1
